requirements.txt
 
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
pyodbc
python-dotenv
